Prevent overwriting of user with team bundle streams
If a job is submitted for a team bundle stream, the
user variable is overwritten when the bundle stream
permissions are checked. Use the cached submitter
variable to test access after the bundle stream is
checked.

Add a unit test for this change and fix a bug revealed
in the test in the handling of group ownership of
device types.

Change-Id: Ic5525e77849f33099c643c22b844a7fecfcc2558