Prevent overwriting of user with team bundle streams
If a job is submitted for a team bundle stream, the
user variable is overwritten when the bundle stream
permissions are checked. Use the cached submitter
variable to test access after the bundle stream is
checked.

Add a unit test for this change and fix a bug revealed
in the test in the handling of group ownership of
device types.

Ported to test_submission.py for django1.6

Change-Id: I9d64650597d0092539a5ee76c29e3bfb67a4163d