Attempt at fixing the vexpress bug where we don't update the uefi
We see a problem quite often on vexpress where, even though we've
updated the uefi, the boot loader doesn't pick it up and flash the uefi
into nor flash. This is an attempt at a simple fix by checking for the
"flashing" message from the bootloader. Multiple retries may get us out
of the hole.

Change-Id: I2e2a164ff587026b63afe7cd944e048d66a0b244