More complete UEFI re-flash fix
Watching on staging I noticed that we're not checking for the flashing
message on test image boot, which we should. I've added it so that we
will. Note that when we change vexpress to a WG type boot commands
device, we'll need to remove the one from _wait_for_master()

Change-Id: I0326d38dc7cea44773f169019fea38f3a8ccdeb5