Prevent leakage of restricted and hidden devices over XMLRPC
Adapt the all_devices and all_device_types API calls
to check for a user in the request. If no user, deny
access to restricted devices and devices of a hidden
device type. If a user is specified, check for devices
to which this user can submit jobs.

Change-Id: I1e6a518c50096dbb2d609e49cd6451420d61ae1e
Closes: #1317127
(cherry picked from commit f560732c7b9df3f28612ce3c0bea3df562bc5fd1)