Preserve SELinux support in test images and tarballs
Allow master image devices to detect if the test image
can retain support for SELinux by checking the version of
tar installed in the master image.

Use SELinux options transparently elsewhere as the dispatcher
is known to support it on all supported LAVA installs.

Drop support if the test image has a version of tar --selinux
outputs "unrecognized" when called in the master image.

Change-Id: I978d7f6cf2c39e4e45f79c517fc48addfd0920da