Remove insert_column workaround
LAVA-1601 - django-tables2 has migrated to OrderedDict
from the deprecated SortedDict. This means that filters
need to create columns in the __init__ in the order in
which the columns are to be displayed, rather than re-ordering
later.

To manipulate the base columns safely, all columns in such
tables need to be instance variables, not class attributes.

Change-Id: I9921228571733e0605c52d79740903adad3d0d3d