Pipeline: fix random failures in the test suite
lava_dispatcher.pipeline.test.test_defs.TestDefinitionHandlers.test_overlay is
failing on my machine because the elements in the list are not in the same
order.
See the Traceback:
  File "lava_dispatcher/pipeline/test/test_defs.py", line 112, in test_overlay
    self.assertEqual(check_list, script_list)
AssertionError: Lists differ: ['lava-vm-groups-setup-host', ... != ['lava-test-case', 'lava-add-k...

Using assertItemsEqual (python >= 2.7) allows to check that both list contains
the same elements, even if the order is different.

Change-Id: I9345e19e142fcd7a389279300240d36541216296