deploy: Add boot_part and root_part parameters
The boot partition and root partition numbers are really properties
of the image to deploy. While it's not uncommon for devices to have a
convention for a common partition layout, that doesn't mean this will be
the default layout for all images.

Add parameters to the deploy_image command to override the defaults if
necessary.

Change-Id: I0fce5694b9bed1197963718b8402a6d09c086d7f
Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>