master: Don't test being in the master image if the serial is down
If the serial connection to a device has been dropped, prodding it to
see if it's in the master image is a bit pointless. Not just pointless,
it also an exception to be thrown by sendline.

So instead, simply assume the device needs to be booted into master
image mode again if the serial connection is gone.

Change-Id: I1dd3f24afe8d017468c8e25dada2112567a4b2fe
Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>