Fix 'Too many open files' error
This error comes from the way logging handler are created. When the pipeline is
created, all handlers are created at the same time. Each of this handler needs
a file descriptor on each log file.
We now create the handlers only before the call to Action.run() and we close
the handler right after.
This decrease the number of open file descriptor but mean that only the run()
functions can log anything using self.logger.

Change-Id: I138fafe3515f6e1dd34ec915fc2b081c46b85265