Master: make DB request partially atomic
Creating a transaction (transaction.atomic) and locking the right DB fields
(select_for_update) is sufficient.
However selecting the actual_device at the same time is not supported by PGSQL
because this field can be NULL.
We the current code, the job is update atomically while the actual_device is
not (no locking).

Change-Id: I1f961c439f84c1b9f59939cfaa492880d4a822e1