lave-test-shell: Add a final test result reflecting the test runs return code
Recently we had an issue with a test in which one step exited
non-zero, causing run.sh to exit-non zero. Unfortunately this isn't
clearly exposed in the tests result bundle view, which simply shows the
pass/failures for the tests in the steps that did run. The only way to
discover there was an issue (apart from double-checking the amount of
test cases ran) is to look at either the log or the return_code
attachment.

Prevent this confusion by adding a synthetic test case to each test run
which only passes if the return_code is 0. This ensures that a test run
exploding has at least one failed result.

Change-Id: If6507480105a59692c591845c550d0b742d8045a