Slave: no need to retry on END messages
As we are using ZMQ, we know that messages will be delivered to the master.
When receiving the message, the master will mark the job as finished in the
database.
The only reason for the master to fail to mark the job is because of a crash
between the reception of the message and the save into the database.
In this case, the master will send a status message when restarting.

Change-Id: I8e4a826a17cb353df23622983a5aeb93e7145571