vmgroups: fix regression with broader exception handler
The issue this patch fixes can be observed on any VMGroups job.
When the VM is asked to reboot, it closes the ssh connection used
for the VM so quickly that pexpect seems to have trouble dealing
with that. The _soft_reboot function will throw an Exception however
the error message is simply the pexpect buffer, in this case '^Creboot
The only way I've found to deal with this issue is to catch the
error with a broader exception handler.

Change-Id: I039f934e07bc2b4c276e35d279ebaf3fee2f78c8