Jobs go incomplete if errors result in no device.
Warn in the logs as this should be picked up
elsewhere, however, ensure that the dispatcher
master continues running in case of this error.
Add a failure_comment to the job if an error is
detected so that the admin can fix the issue.
Support retrieving the job_status, when sent.
Add support for using local templates in unit tests.

Change-Id: I2ee3120fb5d8fd7b7f237ce3eebbdf07abe7478c