lava-test-shell: Add test result for installation
Currently lava-test-shell for each test run, first runs all the install
steps of the various tests followed by running each test one by one.
Unfortunately when the install steps for *one* of the test fails the
whole test run is aborted and no bundle are posted. Meaning that there
is no indication that anything went wrong.

Fix this by:
* Run the install steps for each test right before running the test
* saving the exit code of the install steps
* continue with the with the next step in a run if the installation fails
* Add a first synthetic test case for running the install step

Note that this has a slight semantic change in that each test will now
only run with the dependencies installed by itself and earlier tests in
a test run, rather then running with the dependencies in *all* tests of
a run.

Change-Id: I9c54cc2d642ac5ad688abe9da721c935e4babbb1