Remove existing tests when doing dummy_deploy
When we do dummy_deploy on a fastboot device, we don't remove the old
tests installed on the device, so we get a device that fills up over
time. All this change does is ensure, at the start of the test, that any
existing test defs are removed. Tested on my home setup with a Samsung
note.

Change-Id: I6fc8d81db4b7346c1780860007caf5e5fd1acddd