Fix multiple boot problems on Juno
The new firmware fix for Junos had a bug, whereby it didn't
differentiate between master image firmware and test image firmware.
This caused problems when we did multiple test image/master image boots
because it lost the idea of where its test image firmware was. I've
seperated out the two types so they are held in different paths in the
test directory hierarchy. Tested on playground:
http://playground.validation.linaro.org/scheduler/job/7799/log_file

Change-Id: I8a4aefb514e80a1f43c9f3b4168e426e21ed882b