Workaround for setuptools bug
See http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=786977

The dist directory is already cleared prior to the tarball build.
If the version does not match what is expected, force the name
of the tarball in dist/ to be the one we require and warn of doing so.
Also move the unpacked directory to the one required.

Change-Id: I6bd301961ebd7971dfc1442f94c3f091f8e46393