Expose current_job information
Add column to the admin device view page to give details
of the current job for any device.
Add information to the device type overview table to
indicate the status of the current job for any device.

Helps identify devices which are in an inconsistent state
with respect to the current job.

Also add a list filter to exclude all Retired devices
in the Devices view of the admin interface.

Change-Id: Ib96e0bebf09d158a1563e0d168a4cd1cdc1210d3