Bug 1614 - reserving devices with a current job
Refuse to reserve job for device if idle and has a current_job
Warn in the log files and leave for admin intervention.

Only set Canceled for jobs which are not Complete or Incomplete.
(This clarifies the job state of multinode groups as the final
state for other jobs in the group is not overridden.
The failure comment is retained for all jobs in the group.)

Add the id field to device status transition tables and order
by -id, not created date.

Change-Id: Ief3c81aa34dad8e2d752b5b4f83dcf494831a85b