Add offset parameter to the job_output API method
This commit adds a new parameter to the job_output XMLRPC method
call for specifying the offset in bytes from which to start
reading the output file.

This allows fetching just part of the output file at a time and
can be useful to implement features like 'live' output logs from
other tools.

Change-Id: I8608e704c896977fca86024043585d9e53bd7408