Simplify testset handling
Along with the testset logging changes in the dispatcher, the
dispatcher-master can simplify the handling of the reported
testcases and testsets.
Also, change from testsuite to test_definition as the two are
directly equivalent and the duplication only confuses the issues.
Add initial view for testcases in a testset.
Add documentation on using TestSet.

Change-Id: Ic330354fd7712e079695ab6904977687e4fd2a2f