Loop on logcat rather than waiting for homescreen message
Waiting for the homescreen we used to do a "logcat -s" and then wait
until we saw the message or timed out, and then issued a ctrl c.
The trouble with this is that when we are using an ssh connection
for adb the ctrl c in the android shell terminates the ssh session.
So now use logcat -sd, in a loop.

Change-Id: I96eb85b8f585781f73e8b026e2f2a7396ffc391f
(cherry picked from commit 448b5cb1481f6db5c3b649198d08afcdd3a17e81)