Add support for the compatibility API
The value of the compatibility level is set by the lava-dispatcher
Strategy classes, so can be bumped in the code if the
implementation of pipeline actions used by one or more Strategy
classes changes significantly. Slaves read the compatibility
and compare with the value calculated when parsing the job
pipeline.
Update some master messages for easier administration.
Add pipeline_compatibility to the model to avoid updating the
definition which would drop the comments.
Add to the unit tests, subject to pending changes in the
dispatcher master.

Change-Id: Ifb01a7cfeb3ec7aea4d08d0bbfba6a6b01071ddc