Read job compatibility level from the definition
API compatibility is job-specific, so read the compatibility level
from the job definition (added by the master) and check against
the actual job calculation on the slave.
Set compatibility = 1 for all existing Strategy classes.
Ensure that the PID is recorded properly upon closure as lazy
logging only logs a %s.

Change-Id: I7bfae915af93959f91a16d8461f5dd21b5f2813e