add new pre boot and pre power commands
There are times when you need to control an external piece of hardware
before hard reset and before monitoring the boot. One example of this would
be HiKey, as it requires a relay inline to control the voltage supplied to
the board over the OTG usb port. This is done to allow devices to probe
which are connected to the type a usb ports.

pre_os_cmd is executed directly after boot and pre_power_cmd is executed
before the power is cycled.

Change-Id: I174fd47000e13527dd9a43d264e92ff90c18cf5b