LAVA-333: Check list of supplied prompts for likely errors
Prompts set by the test writer might not contain a colon or other
characters commonly found in actual prompts, causing boot message output
to be wrongly matched as if it was a prompt. e.g. systemd with
openssh-server installed often outputs <user>@<hostname> which would
easily match a test writer prompt value of "root@linaro" - instead this
is reported as a warning in the test logs.

Change-Id: I6e32bdec00dafe8cce734c44bbbcea1ad1e4dfa8
Signed-off-by: Vitaly Chernooky <vitalii.chernookyi@linaro.org>