Fix up network settings
Remove %(lava_network_info)s from console definition, it's nothing to
do with the console!
Add a new ${network} variable, pulling in the $(lava_network_info)s
setting and the ip=dhcp setting.
Also switch from (broken) ip=eth0:dhcp to just
ip=:::::eth0:dhcp. Attempt to specify the first interface as a
default, and expect the initramfs to fix this up later if that's
needed on multi-homed hosts.

Change-Id: Idc5f659a71b0ebf23d01f60b003c682c1daab978