LAVA-507 - Fix assumptions in call_protocols
Jobs can use more than one protocol but any one action is
likely to only have anything to do on a single protocol, so
allow call_protocols to continue to the next protocol in the job.
Not all protocols need to return a reply to a message, an ACK is
sufficient, so allow for these calls as well.

Change-Id: Ibeac8ade214808d520b37bccf61a7d5d84d4115b