Fix handling of ramdisk when combined with NFS
The ordering of the checks caused the overlay and modules
to only be applied to the ramdisk if NFS was specified as well.
Modular kernels may need an initramfs which in turn needs modules
but the test actually happens in the NFS.
Add modules to initramfs and NFS if both are specified.
Add overlay to NFS if initramfs and NFS are both specified.
Tidy up some pylint issues in apply_overlay.py

Change-Id: Ifa1d5eb32e1933c4f8749d4c38225bf646507316