Enforce test_params constraint
Allow deserialization of bundles to continue even if the job uses
a set of test definitions parameters that would cause the bundle to
retrieve a software_context test_params string longer than 1024.
test_params is only used in the BundleDetailTable to represent the
name of the test and is shortened further in the template. The full
list of parameters is preserved in the job definition.
Fixes TICKET-258.

Change-Id: Ib576b32c8742f90a256d4b75b1a77ffb835925b8