Dont use "-p port" in SCP options
The example ssh device tried to pass a -p and -l argument in the ssh
"options" which is used to generate the command parameters.
These options are different between SCP and SSH.
Instead, there was already code to create the correct arguement for
setting the port, so use that instead.
Also, enforce strict host key checking is off to stop scp prompting
for yes/no when first connecting to a device

Change-Id: I083eef0ea916e2330b446e575d737361abe1b3e0