Fix server error if device has no current job
If a device is in an error state where it is Running with no
current job, prevent a HTTP 500 and output a message to the
device status column.
Add logging so that this error and the equivalent in Reserved
state are logged to lava_scheduler_app, not merely reported to
the user.
Add pylint fixes.

Change-Id: Iea0b4295d4192af19e70798ce4ad8d00c2d08948