device-types: add sun5i-r8-chip.conf
This adds device type configuration for the NextThing Co. CHIP
(sun5i-r8-chip) which is U-Boot based.

The Ethernet is provided by a USB to Ethernet adapter with a modified
version of U-Boot. It has to detect USB devices before trying DHCP and
assign an Ethernet interface.
To have Ethernet over USB working in U-Boot, you have to add the
following lines to one of the header used to create the U-Boot version
specific to your board (can be sunxi-common.h for example):
 #define CONFIG_USB_HOST_ETHER
 #define CONFIG_USB_ETHER_ASIX
Check NET, PING, DCHP and USB_EHCI_HCD in menuconfig.

The Ethernet interface is automatically set to asx0 (the USB to Ethernet
adapter).

Change-Id: I4dfd390b78c18b711ee0afaa92c13c5ca5f221d2
Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>