LAVA-477 - migrate option_list to argparse
Use the existing support for django < 1.8 to map argparse
arguments back to option_list whilst covering removal of
optparse support in django1.10 and later.
Change schedulermonitor to allow for argparse remainders
and move the positional arguments to the end of the command
so that the -l and -f arguments are picked up.

Change-Id: If4a92e7ad8160873335f63cbcdfa2ede50631dc5