kvm-aarch64: expose a virtual random number generator (RNG) to the guest
In order to support kernel address space layout randomization (KASLR) on
arm64, expose a virtio-rng device to the guest. This will be used by the
UEFI firmware to expose the EFI_RNG_PROTOCOL to the UEFI stub in the
kernel, which configures KASLR at early boot time.

When doing non-UEFI boots, this device will be simply ignored (unless the
kernel happens to be built with support for virtio-rng over PCI, in which
case it will be wired up to the guest's /dev/hwrng)

Change-Id: Id734c84b9583b0e4281f98d7e772ee5eddb61220
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>