kvm-aarch32: introduce device type for 32-bit guest on 64-bit KVM host
This adds a device type 'kvm-aarch32' based on 'kvm-aarch64' which can be
used to execute 32-bit guests under a 64-bit KVM host. The main difference
with kvm-arm is that it emulates the 'virt' machine rather than a Versatile
express.

The primary purpose is for testing 32-bit guest booting under UEFI, which
is tightly coupled to QEMU's 'virt' machine model. This could potentially
execute under a 32-bit KVM host as well, but it is actually preferable to
run 32-bit and 64-bit guests under the same hypervisor implementation,
since it limits the scope of testing to what runs inside the guest.

Change-Id: If11605197d5fceaa59459d725b5b1c48f43ee8a9
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>