Fix calling the Bzr unit tests
In a VM, the bzr tests would fail - this turned out to be a missing
entry for `bzr whoami` but running the tests with the environment
containing BZR_HOME: /dev/null causes bzr to fail when it cannot
generate a username for the commit and store that config.
Use self.tempdir which is cleaned up at the end of the test.
Add test_utils back to python setup.py test.

Change-Id: I33449a03d6a37a49849e643d2adc6271b35badd9