device-types: add sun8i-a33-sinlinx-sina33.conf
This adds device type configuration for the Sinlinx SinA33
(sun8i-a33-sinlinx-sina33) which is U-Boot based.

The Ethernet is provided by a USB to Ethernet adapter with a modified
version of U-Boot. U-Boot version has to be at least v2016.05-rc1 to be
able to use the adapter.
To have Ethernet over USB working in U-Boot, you have to add the
following lines to one of the header used to create the U-Boot version
specific to your board (can be sunxi-common.h for example):
 #define CONFIG_USB_HOST_ETHER
 #define CONFIG_USB_ETHER_ASIX
 Check NET, PING, DCHP and USB_EHCI_HCD in menuconfig.

Change-Id: I6c993bf3c99c842cb2d4740ad04cdfe3f4c68c14
Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>