lava_scheduler_app: api: Add put_into_online_mode
Add the put_into_online_mode endpoint to the API to put a given device
in online mode with the given reason and skip health check if asked.

To access this endpoint, you have to be authenticated as a user with
sufficient permission (the same as the one used in the web interface:
superuser, device owner or 'can change device' permission).

Skipping health check is optional.

Change-Id: I8b6ffb3b2da00d3ce8940ddb96793726e981a35d
Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>