lava_scheduler_app: api: Add put_into_maintenance_mode
Add the put_into_maintenance_mode endpoint to the API to put a given
device in maintenance mode with the given reason and optionally the
mail address to optionally notify, if a job is currently running, when
it has finished.

To access this endpoint, you have to be authenticated as a user with
sufficient permission (the same as the one used in the web interface:
superuser, device owner or 'can change device' permission).

The mail address to notify is optional.

Change-Id: I6e5b8e7671f405a56286e0626e1e05e7efe2fbc6
Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>