lava_scheduler_app: api: Add get_device_status
Add the get_device_status endpoint to the API to get current status
(idle, offline, running, ...) of the given device.
If its status is offline, the date since which it is offline and the
user responsible of this offlining are joined to this response.
If its status is running, the job id of the running job is attached to
this response.

The queried device has to be visible to the user who makes the request.

Change-Id: Ie5c9cba1ebb294e6fe49e7f5d63d7dd1f492b909
Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>