ldap: Support all LDAP group types in django_auth_ldap
To integrate LAVA with an Active Directory LDAP server it was necessary
to add support for the ActiveDirectoryGroupType. With this patch all
LDAP group types can be used in settings.conf. All group types are
extracted from the module using inspect and the value in settings.conf
is verified to be a valid group name before using exec to import the
requested group type.

Change-Id: I57847131bd3c98f7a7ffb1b06751a6d69af54563