Consider lxc-droid device type as a meta / virtual device type.
Implements:

 * LAVA-669 - Convert nexus4 device to virtual lxc-droid device type
 * LAVA-660 - lxc protocol finalize should identify adb device

The actual device type such as nexus4, nexus10, hikey, etc. will be used
instead of calling the device type as lxc-droid which provides a direct
mapping of these devices to their actual device types. Creating a new device
type such as lxc-droid will prove to hide all the devices and we do not have
any control to request for a specific device of type say nexus4 or something
else (this will get impacted when running jobs via lava-server).

The devices are identified as lxc-droid virtual device type by the presence
of the following things:

  * fastboot command
  * adb serial number
  * device_usb_path
  * connection method should have lxc

Change-Id: If66c71baf995d26a2173e8752d0e139e538d5f1a