Force fsck on next master image boot
We have a lot of issues with the Versatile Express platform, particularly
Juno, where the master image needs an fsck and is waiting for user interaction.
The purpose of this fix is to always force an fsck every time the master image
boots. It may seem like overkill, but an fsck typically only takes a few seconds.

Change-Id: I80dd33fa5f02c1db4c95d4cd6243642e790aa103