Extend kernel command overrides.
Allow the kernel loglevel to be overridden on mustang to allow
'kernel_loglevel' to be set in the job context to replace the
default 'debug' value.
Also allow override to update previous earlyprintk to replacement
earlycon argument, depending on the kernel build.
Add base_kernel_args support to allow extra kernel options
to be supplied.

Change-Id: I8bda2b667f061c846115e560272386473d12f31e