Ensure custom permissions are created
Default permissions are created for add, change or delete
but combined and custom permissions like 'cancel_resubmit_testjob'
need to be created as part of the model Meta class now that
we no longer use south. If the permission already exists, nothing
changes. Add unit test to ensure the change is retained.

Change-Id: Id5206cb3a5c1a18d5e00c20096f58d86dd7dd45d