Unavailable qemu command should not fail unit tests.
There are multiple unit tests which tests the structure of test job and
test definition which does not need 'qemu-system-x86_64', since we are
not running the job itself, but they fail due to unavailability of
'qemu-system-x86_64'. These infrastructure errors can be ignored in
order to continue with more useful things the unit test is actually
testing.

Extend to check the infrastructure error handling for specific
paths.

Initially uncovered from https://ci.debian.net/

Change-Id: I35b357d90fc910b778ec3e7440a10dd90cbdfcac