Fix documentation for test definition name handling
The metadata name inside the YAML is now ignored as the
job submission test definition name is required. The requirement
is that the job can detect when test definitions have not been
executed, using the pipeline level ensures that repeat support
does not interfere with the calculation and provides unique
test suite names.

Change-Id: Ia6d3895ecac0dad8e50313e4d3400513e6a91275