device-types: add armada-385-db-ap.conf
This adds device type configuration for the Marvell DB-A380-AP
(armada-385-db-ap) which is U-Boot based.

This boards boots only with appended DTB and its serial is not the default
ttyS0 but ttyS1.

the ip argument for the kernel command line has been omitted because it
makes the Linux kernel hangs when initializing network interfaces.

Change-Id: I703a6cc9f4a1788f4db6c5853b16cea74e92a744
Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>