device-types: add alpine-db.conf
This adds device type configuration for the Annapurna Labs Alpine Dev Board
(alpine-db) which is U-Boot based.

The following code parts are used to define the clock-frequency of the serial in
Linux kernel Device Tree:

'fdt addr ${fdt_addr_r}; fdt resize; fdt set /soc/uart@fd883000 clock-frequency <375000000>'

The Device Tree has the clock frequency missing and in some older U-Boot
versions, the clock frequency for the serial was not initialized by the
bootloader. A patch has been sent to the Linux kernel upstream.
This maintains compatibiliy for older bootloaders and does not break newer
versions nor it will break when the patch will be merged to Linux kernel
mainline.

The Ethernet port to be used is #A.

Change-Id: I7e906244ad559705016182673dda09dcda80e6bc
Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>