LAVA-200 - track admin actions made in the UI
Create a django admin log record for operations made through the
UI including restrict and derestrict, put online, put offline,
put into looping and cancel reserved so as to also log XML-RPC.
Add explicit logging for updating the device dictionary.
Note that admin operations in the developer shell cannot be tracked.
Remove the error in restrict and derestrict which fails due to
the attempt to create a device state transition from and to the
same state.

Change-Id: Id5f8efb9078d20fbebfdcbce9919c66940ac58ca