LAVA-740 - export vlan names and interface names
Add lava-vland-names which outputs the user-specified vlan name
and the interface label set by the admin which can be used to
correlate with the sysfs/MAC details (also referencing the interface
label) to allow test writers to know which interfaces are on which
vlan. The interface label does not necessarily match the interface
name as this can be changed in userspace.

Also handle case when self.tags in None
Existing for loop splits None into ['N', 'o', 'n', 'e']

Change-Id: I587ede6860671374268392e161ed4614504893c1