Simplify and clarify the "first install" docs
Replace the architecture overview graphic with something closer to
reality, and include svg source too so we can edit it sanely.
Add mention of the scheduler on the master.
More consistent use of terminology, replace "slave" with "worker"
where appropriate.
Move some of the text out into other doc areas where it's going to fit
better and not intimidate new users.

Change-Id: Ic649a262ceef15a14e7f585326ec1c2e78cd35b4