LAVA-748 ensure test definition names are unique
For compatibility with V1 and due to the way that repeats
currently work, the test definition name in the job submission
needs to be unique for the Test Action. If the same definition is
re-used with different parameters, the test definition name needs
to be modified. This does not affect the Lava Test Shell
Definition YAML itself.

Change-Id: I80c43eb21c509429ccb466c7fac955ab6892ea0b