LAVA-765 V1 MultiNode sub_id correction
'latest' is vulnerable to load on the server, causing jobs to create
a sub_id based on the wrong parent_id. Create the .0 sub_id job in
isolation and then create subsequent jobs in the multinode group
directly using the retrieved parent_id. This is likely to increase
the apparent distance between the job.id and job.sub_id but the
templates already handle this.

Change-Id: Ibe55ca3c883fbaa7170b042f0aa60f6adb9d532f