Support parallel (pbzip2) compressed bzip2 files
The python bzip2 decompression library throws an EOFError
if it reaches the end of a bzip2 stream before reaching the end
of a chunk. This is usually caused by a parallel (pbzip2) compressed
files which consist of lots of bzip2 streams concatenated together
into a single file.

Attempt to detect this situation and then continue to decompress
the rest of the data, otherwise throw a legitimate JobError for
other issues.

Change-Id: I8c124aa4daa6db24ce53dcebf617a0ddbec25b41