Add support for the depthcharge boot method
Depthcharge is a Coreboot bootloader payload used on Chrome OS
devices.  This is to add the `depthcharge' boot method with basic
functionality to accept a pre-packaged FIT kernel image file and boot
it.  The kernel command line is stored in a plain file which is then
downloaded alongside the kernel image over TFTP.

This relies on the Depthcharge command line interface to be enabled and
the `tftpboot' command to be available.

Tested on rk3288-veyron-jaq with a single ramdisk and with NFS.

Change-Id: I8c6e588ea4027b9377a7033ff19a747b304fa8ae
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>