Allow install-steps to use the git-repo directly
Tests using install: steps with git-repos: need to be able to
access the repository in the current working directory.
Ensure that git-repos are cloned into the correct sub-directory
and create a repo/ directory for repo.git reference.
Fix some unreachable code and ensure that the connection is
always returned from run()

Change-Id: I7e035a63a8e38c557e1af5f033982e47779d0fbc