master: expect a protocol version with HELLO messages
This is a *light* protection against protocol mismatch but this is not perfect.

In fact, if only the master restart, the slave is not required to send a new
HELLO message, hence the protocol version is not checked. This check will only
happen when a slave restart.

The slaves won't know why the connection is refused and will send HELLO_RETRY
forever. However, the master prints the error in his log.

Change-Id: I4a54a32fccf35c35ab0d5955ffdf91135be0f22a