Give namespaces to all actions and support simpler code flow
- All actions are given the namespace 'common' unless defined
in the job yaml
- 'test_info' parameter is built by the parser and passed to
all actions
- 'test_info' is used to determine if a test type:
needs an overlay
has a shell
needs deployment_data

and all actions have class methods to probe these for the test
actions within their namespace

- boot and deploy actions can use logic in populate() or anywhere
else to decide what other actions to add to the pipeline. This makes
pipelines smaller and jobs run faster

- classes used for 'monitor' boot and deploy types have been
removed as they are no longer needed. jobs can use the normal
boot and deploy classes as they are now aware of the test action
in their namespace (if there is one)

- we no longer build overlays when there is no test action in the job
as there is no need to

- change the accepts() functions in various boot/deploy classes
to accept 'type: monitor' for backwards compatibility

Change-Id: I7305a04fc5014fffa22a8f4970c1f2f46eb4203c