Add the commit-id to the lava testdef results
Make the commit id of the test definition easier to access by adding
to the results of the testdef action.
Make the checksums available in the lava results and thereby include
results for the duration of each download action.
Add the repository, revision and path for direct lookups.
Fix missing import of LAVABug

Change-Id: If63d0cbbea5d7be1a3903ec2ca89cc489230d02a