Prevent intermittent unit test failure.
Assert a fixed order with the retired device at the end to correct
the calculation of how many devices were allowed before the retired
device was encountered.
Also move files from snapshots to a permanent home on images.v.l.o
Enforce that devices are removed at the end of each test.

Change-Id: Id6b080ff50c31049d3dc1219b875d3e2bf86db79