Skip retired and obsolete device types
If all devices of a DeviceType are retired, do not create a health
check file but allow an existing one to be cleaned (it could easily
be invalid syntax).
If the device type is not displayed in the UI (as distinct from
owners_only which only hides from certain users), skip that device
type. Allow cleaning of these device types as well.

Change-Id: I24ba15b72afad6d6ac34332677c906edb55b2e76