Allow LXC protocol to call pre-os-command
LAVA-921 : If LXC is in use, disable the automatic calling of
pre-os-command and use the protocol to give greater flexibility
on when and if the command is called. This allows test writers
to run tests with and without the network connection on HiKey
and allows the OTG port to remain in use during AOSP tests.
Disable some pylint warnings in files included in the change.

Change-Id: I70a2382b600e05f12621b71b699978877251c6ea