Fix scheduling when putting a device into looping
Without this patch, it's possible to get two jobs running on the same device.
Steps to reproduce:
* get an offline device with health in [unknow, fail]
* put it online => the scheduler starts a health check
* before the end of the job, put the device offline and then in looping mode

The device is now still running the health check but with an "Idle" status. A
second health check is then automatically created and assigned to this device.

Allowing only Offline devices to got into looping mode fixes this bug.

Change-Id: I37e8ffeab08bb08997cbcd44b9f5d93441c8fd18