templates: remove duplicated blocks
When a block is defined in the base template, it will be present in the
generated file even if the child template does not reference it.

So adding "{% block foo %}{{ super() }}{% endblock %}" will duplicate the block
"foo" in the generated file.

This duplication is not really an issue due to the way the yaml parser is
handling duplicated dictionaries (the last one overrides the previous ones).

Change-Id: Ifb266e6135871ea96cc09e9d84fd4d70b52506a7