Run login_commands if provided in auto_login
If some login commands are provided, run them in AutoLoginAction
before setting the shell prompt.  A typical use-case for this is to
first log in with a regular user and then switch to root using "sudo
su" or "su".  Doing this changes the prompt, so it can't be done later
otherwise the job would stop waiting for the default prompt to appear.

This change relies on a matching change in the lava-server schema to
accept the login_commands parameter.

Also check in validate() that login_commands parameters are valid and
add a test case in the test_auto_login unit test.

Change-Id: Ibb1c5a441c4621cc08178d7adbe5ba503ed1abec
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>